Move functions dependent on oidc external libraries to lib/services.
