+++ api/types/oidc.go
@@ -24,7 +24,6 @@ import (
 	"github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/utils"
 
-	"github.com/coreos/go-oidc/jose"
 	"github.com/gravitational/trace"
 )
 
@@ -416,15 +415,6 @@ type OIDCConnectorSpecV2 struct {
 	GoogleAdminEmail string `json:"google_admin_email,omitempty"`
 }
 
-// GetClaimNames returns a list of claim names from the claim values
-func GetClaimNames(claims jose.Claims) []string {
-	var out []string
-	for claim := range claims {
-		out = append(out, claim)
-	}
-	return out
-}
-
 // ClaimMapping is OIDC claim mapping that maps
 // claim name to teleport roles
 type ClaimMapping struct {
@@ -434,25 +424,6 @@ type ClaimMapping struct {
 	Value string `json:"value"`
 	// Roles is a list of static teleport roles to match.
 	Roles []string `json:"roles,omitempty"`
-}
-
-// OIDCClaimsToTraits converts OIDC-style claims into the standardized
-// teleport trait format.
-func OIDCClaimsToTraits(claims jose.Claims) map[string][]string {
-	traits := make(map[string][]string)
-
-	for claimName := range claims {
-		claimValue, ok, _ := claims.StringClaim(claimName)
-		if ok {
-			traits[claimName] = []string{claimValue}
-		}
-		claimValues, ok, _ := claims.StringsClaim(claimName)
-		if ok {
-			traits[claimName] = claimValues
-		}
-	}
-
-	return traits
 }
 
 // OIDCConnectorSpecV2Schema is a JSON Schema for OIDC Connector
