+++ lib/services/types.go
@@ -270,9 +270,6 @@ var (
 	NewOIDCConnector          = types.NewOIDCConnector
 	SetOIDCConnectorMarshaler = types.SetOIDCConnectorMarshaler
 	GetOIDCConnectorMarshaler = types.GetOIDCConnectorMarshaler
-
-	GetClaimNames      = types.GetClaimNames
-	OIDCClaimsToTraits = types.OIDCClaimsToTraits
 )
 
 // plugin_data.go
